+++ CHANGELOG.md
@@ -4,6 +4,18 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 Until it reaches 1.0.0, breaking changes will bump the _MINOR_ digit (the second one).
 
+## Unreleased (202x-xx-xx)
+
+Compare with [last version](https://github.com/meduzen/datetime-attribute/compare/0.2.1...main).
+
+### New
+
+- Add `datetime()` missing global precisions.
+
+### Changed
+
+- `datetime()`: `local` and `global` keywords are now `datetime` and `utc`.
+
 ## v0.2.1 (2021-04-02)
 
 Compare with [previous version](https://github.com/meduzen/datetime-attribute/compare/0.1.7...0.2.1).